There are now CMake options to specify which UIs to compile, and mcd_pcm
compiles with MSVC, so we might as well merge everything back to master.

gens-sdl is not quite done, though. Among other things, I need to revert
some DMA stuff that broke "The Adventures of Batman & Robin" and I need
to add Pico tablet support.

+++ src/libgens/tests/EEPRomI2CTest/EEPRomI2CTest.cpp
@@ -205,7 +205,7 @@ void EEPRomI2CTest::CompareByteArrays(const uint8_t *expected, const uint8_t *ac
 
 	// TODO: Use stringstream instead?
 	const uint8_t *pE = expected, *pA = actual;
-	for (size_t i = 0; i < size; i++, pE++, pA++) {
+	for (unsigned int i = 0; i < size; i++, pE++, pA++) {
 		if (i % 16 == 0) {
 			// New line.
 			if (i > 0) {
